% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_comment.R
\name{add_comment}
\alias{add_comment}
\title{Add a comment}
\usage{
add_comment(comment)
}
\arguments{
\item{comment}{text of comment. Markdown elements are supported.}
}
\description{
Add a comment to a jira ticket
}
\author{
Derek Chiu
}
