% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_issue.R
\name{assign_issue}
\alias{assign_issue}
\title{Assigns an issue to a user}
\usage{
assign_issue(name, issue = NULL)
}
\arguments{
\item{name}{username to assign. Set to \code{NULL} to unassign the issue.}

\item{issue}{the issue key. Defaults to the project directory base name,
assuming the R project is named after the issue key.}
}
\description{
Assign an issue to a user for a JIRA ticket
}
\author{
Derek Chiu
}
