% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_file.R
\name{attach_file}
\alias{attach_file}
\title{Attach a file}
\usage{
attach_file(x, add_date = FALSE)
}
\arguments{
\item{x}{path to file to attach}

\item{add_date}{if \code{TRUE}, appends today's date to the attached file name}
}
\description{
Attach a file to a jira ticket
}
\author{
Derek Chiu
}
