% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_jira.R
\name{authenticate_jira}
\alias{authenticate_jira}
\title{Authenticate JIRA credentials}
\usage{
authenticate_jira(username)
}
\arguments{
\item{username}{jira username}
}
\description{
Store JIRA credentials in system credential store.
}
\details{
Stores JIRA username and password in your system's credential store.
Authentication for all RESTful API calls persist for the same R session.
}
\author{
Derek Chiu
}
