% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_attachment.R
\name{delete_attachment}
\alias{delete_attachment}
\title{Deletes an attachment}
\usage{
delete_attachment(id)
}
\arguments{
\item{id}{the ID of the comment}
}
\description{
Deletes an attachment from an issue
}
\details{
The issue ID is actually not needed, only the attachment ID.
}
\author{
Derek Chiu
}
