% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_comment.R
\name{delete_comment}
\alias{delete_comment}
\title{Deletes a comment}
\usage{
delete_comment(id, issue = NULL)
}
\arguments{
\item{id}{the ID of the comment}

\item{issue}{the issue key. Defaults to the project directory base name,
assuming the R project is named after the issue key.}
}
\description{
Delete a comment on a jira ticket
}
\author{
Derek Chiu
}
