% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comments.R
\name{get_comments}
\alias{get_comments}
\title{Returns a comment}
\usage{
get_comments(issue = NULL)
}
\arguments{
\item{issue}{the issue key. Defaults to the project directory base name,
assuming the R project is named after the issue key.}
}
\description{
Return all comments for an issue
}
\author{
Derek Chiu
}
