% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_epic_linked_issues.R
\name{get_epic_linked_issues}
\alias{get_epic_linked_issues}
\title{Returns linked issues to an Epic}
\usage{
get_epic_linked_issues(issue = NULL)
}
\arguments{
\item{issue}{the issue key. Defaults to the project directory base name,
assuming the R project is named after the issue key.}
}
\description{
Return all linked issues to an Epic issue.
}
\author{
Derek Chiu
}
