% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_assignee.R
\name{get_issue_assignee}
\alias{get_issue_assignee}
\title{Returns assignee for an issue}
\usage{
get_issue_assignee(issue = NULL)
}
\arguments{
\item{issue}{the issue key. Defaults to the project directory base name,
assuming the R project is named after the issue key.}
}
\description{
Returns the user designated as the assignee for an issue
}
\author{
Derek Chiu
}
