% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_attachments.R
\name{get_issue_attachments}
\alias{get_issue_attachments}
\title{Returns all attachments for an issue}
\usage{
get_issue_attachments(issue = NULL)
}
\arguments{
\item{issue}{the issue key. Defaults to the project directory base name,
assuming the R project is named after the issue key.}
}
\description{
A list of file names for all attachments for an issue. Attachment IDs
are used as the list names.
}
\author{
Derek Chiu
}
