% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_reporter.R
\name{get_issue_reporter}
\alias{get_issue_reporter}
\title{Returns reporter for an issue}
\usage{
get_issue_reporter(issue = NULL)
}
\arguments{
\item{issue}{the issue key. Defaults to the project directory base name,
assuming the R project is named after the issue key.}
}
\description{
Returns the user who created the issue
}
\author{
Derek Chiu
}
