% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_transitions.R
\name{get_issue_transitions}
\alias{get_issue_transitions}
\title{Returns all transitions that can be applied to an issue}
\usage{
get_issue_transitions(issue = NULL)
}
\arguments{
\item{issue}{the issue key. Defaults to the project directory base name,
assuming the R project is named after the issue key.}
}
\description{
A data frame of available transitions to apply to an issue and their ids. The
available transitions change depending on the current issue status.
}
\author{
Derek Chiu
}
