% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_issue_worklogs.R
\name{get_issue_worklogs}
\alias{get_issue_worklogs}
\title{Returns the total time spent on an issue}
\usage{
get_issue_worklogs(issue = NULL)
}
\arguments{
\item{issue}{the issue key. Defaults to the project directory base name,
assuming the R project is named after the issue key.}
}
\description{
The individual worklogs are added up and displayed in a human-friendly
duration format.
}
\author{
Derek Chiu
}
