% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label_worklogs.R
\name{get_label_worklogs}
\alias{get_label_worklogs}
\title{Get time spent on issues with specified labels}
\usage{
get_label_worklogs(
  labels,
  project = "BC",
  from = NULL,
  to = NULL,
  maxResults = 50
)
}
\arguments{
\item{labels}{vector of labels}

\item{project}{jira project name}

\item{from}{start date of range to consider issues. Defaults to first day
of current year}

\item{to}{end date of range to consider issues. Defaults to current date.}

\item{maxResults}{maximum number of results to return. Defaults to 50.}
}
\description{
The duration of time spent on all issues related to a set of labels within a
date range is reported.
}
\author{
Derek Chiu
}
