% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_user_worklogs.R
\name{get_user_worklogs}
\alias{get_user_worklogs}
\title{Get time spent on all issues for a user}
\usage{
get_user_worklogs(user, project = "BC", from = NULL, to = NULL)
}
\arguments{
\item{user}{username}

\item{project}{jira project name}

\item{from}{start date of range to consider issues. Defaults to first day
of current year}

\item{to}{end date of range to consider issues. Defaults to current date.}
}
\description{
The duration of time spent on all issues related to a user within a date
range is reported. Both the issues assigned to a user and issues reported by
a user but unassigned are included.
}
\author{
Derek Chiu
}
