% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jirarest-package.R
\docType{package}
\name{jirarest-package}
\alias{jirarest}
\alias{jirarest-package}
\title{jirarest: Accessing the Jira REST API}
\description{
Access the Jira REST API hosted on the BCGSC. Allows authenticated users to attach files and add comments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TalhoukLab/jirarest}
  \item \url{https://talhouklab.github.io/jirarest/}
  \item Report bugs at \url{https://github.com/TalhoukLab/jirarest/issues}
}

}
\author{
\strong{Maintainer}: Derek Chiu \email{dchiu@bccrc.ca} (\href{https://orcid.org/0000-0002-7591-4881}{ORCID})

Authors:
\itemize{
  \item Derek Chiu \email{dchiu@bccrc.ca} (\href{https://orcid.org/0000-0002-7591-4881}{ORCID})
}

}
\keyword{internal}
