% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_auth.R
\name{store_auth}
\alias{store_auth}
\title{Store JIRA credentials to authenticate with OpenSpecimen API}
\usage{
store_auth(username)
}
\arguments{
\item{username}{jira username}
}
\description{
Store JIRA credentials in system credential store.
}
\details{
Stores JIRA username and password in your system's credential store.
Authentication for all RESTful API calls to OpenSpecimen persist for the same
R session.
}
\author{
Derek Chiu
}
