% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_labels.R
\name{update_labels}
\alias{update_labels}
\title{Update labels on a jira ticket}
\usage{
update_labels(labels_add = NULL, labels_remove = NULL, issue = NULL)
}
\arguments{
\item{labels_add}{labels to add to issue}

\item{labels_remove}{labels to remove from issue}

\item{issue}{the issue key. Defaults to the project directory base name,
assuming the R project is named after the issue key.}
}
\description{
Update the labels for a jira ticket
}
\author{
Derek Chiu
}
