% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_comment.R
\name{get_comment}
\alias{get_comment}
\title{Returns a comment}
\usage{
get_comment(id, issue = NULL)
}
\arguments{
\item{id}{the ID of the comment}

\item{issue}{the issue key. Defaults to the project directory base name,
assuming the R project is named after the issue key.}
}
\description{
Return a single comment of a jira ticket
}
\author{
Derek Chiu
}
